rename register file
