+++ packages/types/src/services/consentUi/requests/index.ts
@@ -17,4 +17,4 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export * from './register.js';
+export * from './RegisterRequest.js';
